+++ src/main/java/mprj/mp/br/calculos/domain/jpa/INPC.java
@@ -16,7 +16,7 @@ public class INPC {
     private String Nome;
 
     @Column(name = "fator")
-    private float fator;
+    private double fator;
 
     @Column(name = "valor")
     private double valor;
@@ -35,7 +35,7 @@ public class INPC {
     public void setValor(double valor){ this.valor = valor;}
     public Date getData() {return data;}
     public void  setData(Date data){ this.data = data;}
-    public float getFator(){return fator;}
+    public double getFator(){return fator;}
     public void setFator(float fator) { this.fator = fator;}
     public String getNome() { return Nome;}
     public void setNome(String Nome){this.Nome = Nome;}
